Implement `open_oneof`
Salvage `prost-derive` into `open-prost-oneof` in order to expose a
new procmacro called `open_openof`.

The impact of this proc macro is to inject a `scalar::OpenVariant`
type inside `Option` types used for oneofs. This new type is defined
externally and has the same structure as a regular option:
```
pub enum OpenVariant<T> {
    Known(T),
    Unknown,
}
```

`open_openof` can be enabled from a `build.rs` file. For example:
```
tonic_build::configure()
  // [...]
  .message_attribute(".", "#[open_prost_oneof::open_oneof()]")
  // [...]
```

`open_oneof` suppresses the `prost::Message` proc macro for the proto
messages (i.e. not the enumerations and oneofs) and generates the same
functions as `prost::Message` with a few diffs for oneof fields in
order to use `OpenVariant`.

Diffs are minimal in order to faciliate rebase on official prost repo.

All the other crates in the repo are neutralized because they require
`prost-derive` which doesn't exist anymore.

--

The current work could be modified to include the encoded bytes inside
the `Unknown` payload.

A similar result for enumerations could be achieved.

--

A variation of this work could be upstreamed to the official prost
implem. It would require:
- Implementing `OpenVariant` somewhere inside `prost` or allow the
user to pass the name of their own type to `prost_build`.
- Changing `prost_build::Config` to add new options in order for the
users to opt-in and out of the `OpenVariant` feature.
- Changing `prost_build` to generate code with `OpenVariant`.
- Changing `prost::Message` to take an `with_openvariant_for_oneof`
parameter.
- Changing `prost_build` to generate such an updated version of
`prost::Message`.
- Upstreaming the diffs in `prost-derive/src/field/oneof.rs`.